added hashed psswrd and endpoints to other folder
