+++ backend/apis/dependencies.py
@@ -1,0 +1,9 @@
+from sqlalchemy.orm import Session
+from database import SessionLocal
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
