+++ backend/auth.py
@@ -1,0 +1,10 @@
+import bcrypt
+
+# Funkcja do haszowania hasła
+def hash_password(password: str) -> str:
+    salt = bcrypt.gensalt()
+    hashed_password = bcrypt.hashpw(password.encode('utf-8'), salt)
+    return hashed_password.decode('utf-8')
+
+def verify_password(stored_hash: str, password: str) -> bool:
+    return bcrypt.checkpw(password.encode('utf-8'), stored_hash.encode('utf-8'))
