+++ backend/requirements.txt
@@ -4,3 +4,4 @@ sqlalchemy
 psycopg2-binary
 pydantic
 email_validator
+bcrypt
